Pull DataFlowScannerOption into SystemOption
