+++ clients/admin-ui/src/features/config-wizard/AddSystem.tsx
@@ -1,71 +1,16 @@
 import { Box, Heading, SimpleGrid, Stack, Text } from "@fidesui/react";
 
-import { useAppDispatch, useAppSelector } from "~/app/hooks";
-import ConnectedCircle from "~/features/common/ConnectedCircle";
-import { useFeatures } from "~/features/common/features";
+import { useAppDispatch } from "~/app/hooks";
 import {
   AWSLogoIcon,
-  DataFlowScannerLogo,
   ManualSetupIcon,
   OktaLogoIcon,
 } from "~/features/common/Icon";
-import { selectDataFlowScannerStatus } from "~/features/plus/plus.slice";
-import { ClusterHealth, ValidTargets } from "~/types/api";
+import { ValidTargets } from "~/types/api";
 
 import { changeStep, setAddSystemsMethod } from "./config-wizard.slice";
-import SystemOption from "./SystemOption";
+import SystemOption, { DataFlowScannerOption } from "./SystemOption";
 import { SystemMethods } from "./types";
-
-const DataFlowScannerOption = ({ onClick }: { onClick: () => void }) => {
-  const { plus, dataFlowScanning } = useFeatures();
-  const scannerStatus = useAppSelector(selectDataFlowScannerStatus);
-
-  const clusterHealth = scannerStatus?.cluster_health ?? "unknown";
-  const isClusterHealthy = clusterHealth === ClusterHealth.HEALTHY;
-
-  // If Plus is not enabled, do not show this feature at all
-  if (!plus) {
-    return null;
-  }
-
-  let tooltip = "";
-  if (!dataFlowScanning) {
-    tooltip =
-      "The data flow scanner is not enabled, please check your configuration.";
-  } else if (!isClusterHealthy) {
-    tooltip = `Your cluster appears not to be healthy. Its status is ${clusterHealth}.`;
-  }
-
-  const disabled = !dataFlowScanning || !isClusterHealthy;
-
-  return (
-    <Box position="relative">
-      <SystemOption
-        label="Data flow scan"
-        description="Automatically discover new systems in your Kubernetes infrastructure"
-        icon={<DataFlowScannerLogo boxSize={8} />}
-        onClick={onClick}
-        disabled={disabled}
-        title={disabled ? tooltip : undefined}
-        data-testid="data-flow-scan-btn"
-      />
-      {dataFlowScanning ? (
-        <ConnectedCircle
-          connected={isClusterHealthy}
-          title={
-            isClusterHealthy
-              ? "Cluster is connected and healthy"
-              : `Cluster is ${clusterHealth}`
-          }
-          position="absolute"
-          right={-1}
-          top={-1}
-          data-testid="cluster-health-indicator"
-        />
-      ) : null}
-    </Box>
-  );
-};
 
 const SectionTitle = ({ children }: { children: string }) => (
   <Heading
