+++ clients/admin-ui/src/features/config-wizard/SystemOption.tsx
@@ -1,5 +1,12 @@
 import { Box, Button, ButtonProps, Text } from "@fidesui/react";
 
+import { useAppSelector } from "~/app/hooks";
+import ConnectedCircle from "~/features/common/ConnectedCircle";
+import { useFeatures } from "~/features/common/features";
+import { DataFlowScannerLogo } from "~/features/common/Icon";
+import { selectDataFlowScannerStatus } from "~/features/plus/plus.slice";
+import { ClusterHealth } from "~/types/api";
+
 const SystemOption = ({
   label,
   description,
@@ -46,5 +53,62 @@ const SystemOption = ({
     </Box>
   </Button>
 );
+
+/**
+ * Wrapper around SystemOption which handles data flow scanner specific
+ * logic, such as cluster health
+ * @param param0
+ * @returns
+ */
+export const DataFlowScannerOption = ({ onClick }: { onClick: () => void }) => {
+  const { plus, dataFlowScanning } = useFeatures();
+  const scannerStatus = useAppSelector(selectDataFlowScannerStatus);
+
+  const clusterHealth = scannerStatus?.cluster_health ?? "unknown";
+  const isClusterHealthy = clusterHealth === ClusterHealth.HEALTHY;
+
+  // If Plus is not enabled, do not show this feature at all
+  if (!plus) {
+    return null;
+  }
+
+  let tooltip = "";
+  if (!dataFlowScanning) {
+    tooltip =
+      "The data flow scanner is not enabled, please check your configuration.";
+  } else if (!isClusterHealthy) {
+    tooltip = `Your cluster appears not to be healthy. Its status is ${clusterHealth}.`;
+  }
+
+  const disabled = !dataFlowScanning || !isClusterHealthy;
+
+  return (
+    <Box position="relative">
+      <SystemOption
+        label="Data flow scan"
+        description="Automatically discover new systems in your Kubernetes infrastructure"
+        icon={<DataFlowScannerLogo boxSize={8} />}
+        onClick={onClick}
+        disabled={disabled}
+        title={disabled ? tooltip : undefined}
+        data-testid="data-flow-scan-btn"
+      />
+      {dataFlowScanning ? (
+        <ConnectedCircle
+          connected={isClusterHealthy}
+          title={
+            isClusterHealthy
+              ? "Cluster is connected and healthy"
+              : `Cluster is ${clusterHealth}`
+          }
+          position="absolute"
+          right={-1}
+          top={-1}
+          data-testid="cluster-health-indicator"
+        />
+      ) : null}
+    </Box>
+  );
+};
 
 export default SystemOption;
